Added onboard LED toggling via setTimeout
Just a little example of playing with only the Arduino, where pin 13
also happens to be the onboard light. This example simply toggles the
light on and off manually with a varying timing based on a stepping
variable and min/max values.

If you went to jsconf, you can run this with only your Arduino, out of
the box. Should be as simple as this:

* Hook up the Arduino to your computer
* Ensure you're on node 0.10.x
* `npm install johnny-five`
* `node onboard-led-toggle.js`

You'll see console messages going by in sync with the little LED
blinkin'. It's spiffy!